+++ test/webtest-commands.ts
@@ -1,11 +1,11 @@
 import {Wordgard} from "wordgard/editor"
-import {Command, deleteLine} from "wordgard/command"
+import {Command, deleteLine, killToLineEnd, yankKilled} from "wordgard/command"
 import {GardState} from "wordgard/state"
 import {Plot} from "wordgard/doc"
 import ist from "ist"
 import {tempEditor} from "./tempview.ts"
 import {basicBuilders, maybeTag, eq} from "./schema.ts"
-const {doc, p} = basicBuilders
+const {doc, p, br, strong} = basicBuilders
 
 function test1<T>(doc: Plot.Doc, cmd: Command<T>, arg: T, expect: Plot.Doc | null, config: GardState.Extension = []) {
   let wg = tempEditor(doc, config)
@@ -31,5 +31,61 @@ describe("deleteLine", () => {
 
   it("doesn't clear past wrap points", () => {
     test(doc(p("bigwordone bigword", 0, "two bigwordthree")), deleteLine, doc(p("bigwordone ", 0, "bigwordthree")), narrow)
+  })
+})
+
+describe("killToLineEnd", () => {
+  it("deletes content after the cursor", () => {
+    test(doc(p("one ", 0, "two")), killToLineEnd, doc(p("one ", 0)))
+  })
+
+  it("joins textblocks when at the end of one", () => {
+    test(doc(p("one", 0), p("two")), killToLineEnd, doc(p("one", 0, "two")))
+  })
+
+  it("deletes line breaks", () => {
+    test(doc(p("one", 0, br, "two")), killToLineEnd, doc(p("one", 0, "two")))
+  })
+
+  it("does nothing at end of doc", () => {
+    test(doc(p("one", 0)), killToLineEnd, null)
+  })
+
+  it("stores content in the kill buffer", () => {
+    let wg = tempEditor(doc(p("a", 0, "b")))
+    Command.dispatch(wg, killToLineEnd)
+    Command.dispatch(wg, yankKilled)
+    Command.dispatch(wg, yankKilled)
+    ist(wg.state.doc, doc(p("abb")), eq)
+    ist(wg.state.selection.head, 4)
+  })
+
+  it("accumulates killed content", () => {
+    let wg = tempEditor(doc(p("a", 0, "b", strong("c")), p("d")))
+    Command.dispatch(wg, killToLineEnd)
+    Command.dispatch(wg, killToLineEnd)
+    Command.dispatch(wg, killToLineEnd)
+    ist(wg.state.doc, doc(p("a")), eq)
+    Command.dispatch(wg, yankKilled)
+    ist(wg.state.doc, doc(p("a", 0, "b", strong("c")), p("d")), eq)
+  })
+
+  it("stops accumulating on cursor motion", () => {
+    let wg = tempEditor(doc(p("a", 0, "b"), p("cd")))
+    Command.dispatch(wg, killToLineEnd)
+    wg.dispatch({selection: {anchor: 5}})
+    Command.dispatch(wg, killToLineEnd)
+    ist(wg.state.doc, doc(p("a"), p("c")), eq)
+    Command.dispatch(wg, yankKilled)
+    ist(wg.state.doc, doc(p("a"), p("cd")), eq)
+  })
+
+  it("stops accumulating on document changes", () => {
+    let wg = tempEditor(doc(p("a", 0, "b"), p("cd")))
+    Command.dispatch(wg, killToLineEnd)
+    wg.dispatch({changes: {from: 5, to: 6}})
+    Command.dispatch(wg, killToLineEnd)
+    Command.dispatch(wg, yankKilled)
+    ist(wg.state.doc, doc(p("a"), p("c")), eq)
   })
 })